requirements.txt
numpy
scikit-learn
pandas
matplotlib
statsmodels
seaborn
python-binance
ccxt
python-dotenv
